+++ lib/capybara/node/matchers.rb
@@ -632,7 +632,7 @@ module Capybara
       #     @option options [Range]   :between (nil)   Range of times that is expected to contain number of times text occurs
       #     @option options [Numeric] :wait (Capybara.default_max_wait_time)      Maximum time that Capybara will wait for text to eq/match given string/regexp argument
       #     @option options [Boolean] :exact (Capybara.exact_text) Whether text must be an exact match or just substring
-      #     @option options [Boolean] :collapse_ws (false) When true replace all whitespace with standard spaces and collapse consecutive whitespace to a single space
+      #     @option options [Boolean] :nomalize_ws (false) When true replace all whitespace with standard spaces and collapse consecutive whitespace to a single space
       #   @overload $0(text, **options)
       #     @param [String, Regexp] text               The string/regexp to check for. If it's a string, text is expected to include it. If it's a regexp, text is expected to match it.
       #     @option options [Integer] :count (nil)     Number of times the text is expected to occur
@@ -641,7 +641,7 @@ module Capybara
       #     @option options [Range]   :between (nil)   Range of times that is expected to contain number of times text occurs
       #     @option options [Numeric] :wait (Capybara.default_max_wait_time)      Maximum time that Capybara will wait for text to eq/match given string/regexp argument
       #     @option options [Boolean] :exact (Capybara.exact_text) Whether text must be an exact match or just substring
-      #     @option options [Boolean] :collapse_ws (false) When true replace all whitespace with standard spaces and collapse consecutive whitespace to a single space
+      #     @option options [Boolean] :normalize_ws (false) When true replace all whitespace with standard spaces and collapse consecutive whitespace to a single space
       # @raise [Capybara::ExpectationNotMet] if the assertion hasn't succeeded during wait time
       # @return [true]
       #