+++ lib/capybara/spec/session/has_text_spec.rb
@@ -36,7 +36,7 @@ Capybara::SpecHelper.spec '#has_text?' do
 
   it 'should search whitespace collapsed text' do
     @session.visit('/with_html')
-    expect(@session).to have_text('text with whitespace', collapse_ws: true)
+    expect(@session).to have_text('text with whitespace', normalize_ws: true)
   end
 
   it 'should be false if the given text is not on the page' do